requirements.txt
apache-airflow
apache-airflow-providers-amazon==7.8.0
pandas==2.2.1
boto3==1.34.52
botocore==1.34.52